general screen added and refactored
